gcf.py, get_gcf.py, fraction_simplify.py
gcf.py is the primary program in this package. The other two programs are CLI applications that import gcf.py to run their primary functions.

get_gcf.py will allow the user to input as many integers as they wish and then will use the getGCF() function in gcf.py to calculate the GCF of all the numbers(even if the GCF=1).

fraction_simplify.py will let the user input a numerator and denominator, assuming the numbers are integers, and will then output a simplified version of that fraction by using the same function to calculate the GCF of the two numbers.